# .kilo/worktrees/equinox-ocean/docs/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "JurisLab API",
  "description": "Multi-agent legal simulation and document review platform (educational).",
  "version": "1.0.0"
 },
 "paths": {
...
  "/api/v1/reviews/{review_id}": {
   "get": {
    "tags": [
     "reviews"
    ],
    "summary": "Get Review",
    "operationId": "get_review_api_v1_reviews__review_id__get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "review_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Review Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Get Review Api V1 Reviews  Review Id  Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/v1/knowledge/search": {
   "post": {
    "tags": [
     "knowledge"
    ],
    "summary": "Search",
    "operationId": "search_api_v1_knowledge_search_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/SearchParams"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "type": "object",
         "title": "Response Search Api V1 Knowledge Search Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   }
  },
  "/api/v1/knowledge/insights": {
   "get": {
    "tags": [
     "knowledge"
    ],
    "summary": "Insights",
    "operationId": "insights_api_v1_knowledge_insights_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "type": "object",
         "title": "Response Insights Api V1 Knowledge Insights Get"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   }
  },
  "/api/v1/knowledge/teaching_pack": {
   "post": {
    "tags": [
     "knowledge"
    ],
    "summary": "Teaching Pack",
    "operationId": "teaching_pack_api_v1_knowledge_teaching_pack_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/SearchParams"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "type": "object",
         "title": "Response Teaching Pack Api V1 Knowledge Teaching Pack Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   }
  },
  "/api/v1/knowledge/collections": {
   "get": {
    "tags": [
     "knowledge"
    ],
    "summary": "List Collections",
    "operationId": "list_collections_api_v1_knowledge_collections_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "type": "object",
         "title": "Response List Collections Api V1 Knowledge Collections Get"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   },
   "post": {
    "tags": [
     "knowledge"
    ],
    "summary": "Create Collection",
    "operationId": "create_collection_api_v1_knowledge_collections_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/CollectionCreate"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "type": "object",
         "title": "Response Create Collection Api V1 Knowledge Collections Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   }
  },
  "/api/v1/knowledge/collections/{collection_id}/items": {
   "post": {
    "tags": [
     "knowledge"
    ],
    "summary": "Add To Collection",
    "operationId": "add_to_collection_api_v1_knowledge_collections__collection_id__items_post",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "collection_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Collection Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/CollectionItemAdd"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Add To Collection Api V1 Knowledge Collections  Collection Id  Items Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/exports/reviews/{review_id}": {
   "get": {
    "tags": [
     "exports"
    ],
    "summary": "Export Review",
    "operationId": "export_review_api_v1_exports_reviews__review_id__get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "review_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Review Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Export Review Api V1 Exports Reviews  Review Id  Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/exports/case_studies/{case_study_id}": {
   "get": {
    "tags": [
     "exports"
    ],
    "summary": "Export Case Study Route",
    "operationId": "export_case_study_route_api_v1_exports_case_studies__case_study_id__get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "case_study_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Case Study Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Export Case Study Route Api V1 Exports Case Studies  Case Study Id  Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "CollectionCreate": {
    "properties": {
     "name": {
      "type": "string",
      "minLength": 1,
      "title": "Name"
     },
     "description": {
      "type": "string",
      "title": "Description",
      "default": ""
     }
    },
    "type": "object",
    "required": [
     "name"
    ],
    "title": "CollectionCreate"
   },
   "CollectionItemAdd": {
    "properties": {
     "item_type": {
      "type": "string",
      "pattern": "^(document|scenario|simulation|review)$",
      "title": "Item Type"
     },
     "item_id": {
      "type": "string",
      "title": "Item Id"
     },
     "note": {
      "type": "string",
      "title": "Note",
      "default": ""
     }
    },
    "type": "object",
    "required": [
     "item_type",
     "item_id"
    ],
    "title": "CollectionItemAdd"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "SearchParams": {
    "properties": {
     "query": {
      "type": "string",
      "title": "Query",
      "default": ""
     },
     "jurisdiction": {
      "type": "string",
      "title": "Jurisdiction",
      "default": ""
     },
     "domain": {
      "type": "string",
      "title": "Domain",
      "default": ""
     },
     "doc_type": {
      "type": "string",
      "title": "Doc Type",
      "default": ""
     },
     "source_status": {
      "type": "string",
      "title": "Source Status",
      "default": ""
     },
     "limit": {
      "type": "integer",
      "title": "Limit",
      "default": 20
     }
    },
    "type": "object",
    "title": "SearchParams"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   }
  },
  "securitySchemes": {
   "HTTPBearer": {
    "type": "http",
    "scheme": "bearer"
   }
  }
 },
...
}